Restore auto group text for players that do not have an authenticator attached to their account (fixes #86)
By overwriting C_LFGList.GetPlaystyleString, we taint the code writing the tooltip (which does not matter), and also code related to the dropdows where you can select the playstyle. The only relevant protected function here is C_LFGList.SetEntryTitle, which is only called from LFGListEntryCreation_SetTitleFromActivityInfo. Players that do not have an authenticator attached to their account cannot set the title or comment when creating groups. Instead, Blizzard sets the title programmatically. If we taint this function, these players can not create groups anymore, so we check on an arbitrary mythic plus dungeon if the player is authenticated to create a group.